requirements.txt
python-dotenv
pymongo
discord.py
dnspython
ruamel.yaml